feat: Add SettingsDataSource to the service locator for settings persistence.

+++ lib/domain/usecases/process_media_usecase.dart
@@ -321,7 +321,7 @@ class ProcessMediaUseCase {
                   width,
                   height,
                   detectionResult.faces,
-                  options.blurStrength,
+                  options.blurStrength.toInt(),
                 );
               }
             } catch (e) {
